Создает компонент Overview и подключает его в компонент Tabs
