+++ src/components/tabs/tabs.test.e2e.test.js
@@ -2,46 +2,25 @@ import React from "react";
 import {configure, shallow} from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
 import Tabs from "./tabs.jsx";
-import {tabs} from "../../test-mocks/test-tabs";
+import {testMovies} from "../../test-mocks/test-films";
 
 configure({
   adapter: new Adapter(),
 });
 
-const activeTab = tabs[1][`type`];
+const movie = testMovies[0];
 
 describe(`Tabs component`, () => {
-  it(`should run callback on Tab link click`, () => {
-    const onClick = jest.fn();
+  it(`should pass Tab name to state.activeTab on link click`, () => {
+    const tabName = `Details`;
     const tabsComponent = shallow(
         <Tabs
-          activeTab={activeTab}
-          onClick={onClick}
-          tabs={tabs}
+          movie={movie}
         />
     );
-    const link = tabsComponent.find(`.movie-nav__link`).at(2);
+    const link = tabsComponent.find(`.movie-nav__link`).at(1);
 
-    link.simulate(`click`, {preventDefault: () => {}});
-    expect(onClick).toHaveBeenCalledTimes(1);
-  });
-
-  it(`should pass tabName to callback on Tab link click`, () => {
-    const tabName = tabs[2][`type`];
-    const onClick = jest.fn((...args) => [...args]);
-
-    const tabsComponent = shallow(
-        <Tabs
-          activeTab={activeTab}
-          onClick={onClick}
-          tabs={tabs}
-        />
-    );
-
-    const link = tabsComponent.find(`.movie-nav__link`).at(2);
-
-    link.simulate(`click`, {preventDefault: () => {}});
-    expect(onClick).toHaveBeenCalledTimes(1);
-    expect(onClick.mock.calls[0][0]).toEqual(tabName);
+    link.simulate(`click`, {preventDefault: () => {}, target: {id: tabName}});
+    expect(tabsComponent.state().activeTab).toEqual(tabName);
   });
 });
